+++ src/config.ts
@@ -22,7 +22,7 @@ export interface IMarpCLIArguments {
 export type IMarpCLIConfig = Overwrite<
   Omit<IMarpCLIArguments, 'configFile'>,
   {
-    engine?: ResolvableEngine
+    engine?: ResolvableEngine | ResolvableEngine[]
     html?: ConverterOption['html']
     lang?: string
     options?: ConverterOption['options']
@@ -45,12 +45,12 @@ export class MarpCLIConfig {
   }
 
   async converterOption(): Promise<ConverterOption> {
-    const engine = await (async () => {
+    const engine = await (() => {
       if (this.args.engine) return resolveEngine(this.args.engine)
       if (this.conf.engine)
         return resolveEngine(this.conf.engine, this.confPath)
 
-      return resolveEngine(Marp)
+      return resolveEngine(['@marp-team/marp-core', Marp])
     })()
 
     const output = this.args.output || this.conf.output